Day4-Django-advanced/django-starter/pyproject.toml
[tool.poetry]
name = "imagine"
version = "0.1.0"
description = "example of imagine autogenerated code"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "*"
django-filter = "*"
django-environ = "*"
drf-yasg = "*"
djangorestframework = "*"
djangorestframework_simplejwt = "5.2.2" 
postgres = "*"
gunicorn = "*"
django-redis = "*"
redis = "*"
python-dotenv = "*"
sentry-sdk = "*"
pycryptodome = "*"
python-dateutil="2.8.2"
celery="*"
flower="*"
requests="*"
pdfkit="*"
bs4="*"
boto3="*"
django-storages="*"
django-ses="*"
django-ratelimit="*"
pyfcm="*"
django-extensions="*"
whitenoise="*"
rapidfuzz="*"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
setuptools = "^68.2.2"
python-decouple = "^3.8"
django-crontab = "^0.7.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.1"
factory-boy = "*"
django-stubs = "*"
mypy = "*"
djangorestframework-stubs = "*"
autopep8 = "*"
flake8 = "*"
isort = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
